oops, more app -> bootstrap renamings
