+++ app/main.mock.js
@@ -16,7 +16,7 @@
 window.onerror = require('./onerror');
 window.config = require('../config.app.js');
 
-var app = window.app = require('./app');
+var app = window.app = require('./bootstrap');
 
 var mock = require('../mock');
 app.useMock(mock);
